Hack to get around GitHub playing with TAR
For whatever reason, GitHub randomly removes the version string from
the root directory of downloaded TAR files.

Try to address this by having both a versioned and an unversioned name
for the root directory, and if either match, accept that.

A fuller solution might add a "unversioned" parameter to checkout() so
we can turn this feature on and off, but that is a task for another day.